STEP 5. People dragon drop.
