--- altitude/static/js/person/dragon-drop.js
@@ -1,13 +1,0 @@
-import {
-    dragged,
-    dragMoveListener,
-    setFixedPositionWhileDragging,
-} from "../common/dragon-drop.js"
-
-interact("#person .drag-drop").draggable({
-    listeners: {
-        move: dragMoveListener,
-        start: setFixedPositionWhileDragging,
-        end: dragged,
-    },
-})
