+++ altitude/views/AGENTS.md
@@ -29,14 +29,12 @@ controllers as `"<!doctype html>" + template(...)`. They regularly include inlin
 |---|---|
 | `js/folders/` | `htmx/folders.scala.html`, `htmx/folder_children.scala.html` (drag-and-drop only) |
 | `js/search-results/` | `includes/search_results.scala.html`, `htmx/results_grid.scala.html` |
-| `js/people/` | `htmx/people.scala.html`, `htmx/person.scala.html`, `htmx/person_inner.scala.html` (drag-and-drop only) |
-| `js/person/` | `htmx/person.scala.html` (single person view) |
-| `js/frontend-app.js` | app-wide bootstrap, shared Alpine stores, search-results hydration, asset actions, people/trash event handling, batch drag wiring, folder event/HTMX handling |
+| `js/frontend-app.js` | app-wide bootstrap, shared Alpine stores, search-results hydration, asset actions, people/person drag-and-drop, people/trash event handling, batch drag wiring, folder event/HTMX handling |
 | `js/common/` | shared: modal, snackbar, navigation, nodes |
 | `js/models/folder.js` | DOM wrapper for folder tree elements |
 | `js/alpine/components/selectable.js` | asset grid multi-select |
 
-Each JS section has a consistent file split:
+Legacy feature folders often used a consistent file split:
 - `event_handlers.js` — custom DOM event listeners (`document.body.addEventListener(Const.events.*)`)
 - `htmx_event_handlers.js` — HTMX lifecycle listeners (`htmx:beforeRequest`, `htmx:afterRequest`)
 - `dragon-drop.js` — interact.js drag-and-drop wiring (except batch drag wiring, which now lives in `js/frontend-app.js`)
@@ -82,8 +80,9 @@ Two-layer event bus, both on `document.body`:
 Batch ops escalate a single-asset drag to a batch when selected assets exist: the
 `assetMoved`/`assetTrashed` handlers are now registered centrally in `js/frontend-app.js`,
 which re-dispatches `batchAssetsMoved`/`batchAssetsRecycled` if the
-`selectedAssets` store is non-empty. Person merge/name/cover-face events and
-trash purge request outcomes are also handled centrally there.
+`selectedAssets` store is non-empty. Person merge/name/cover-face events,
+person discard follow-up, people/person drag-and-drop, and trash purge request
+outcomes are also handled centrally there.
 
 ### Full action flow (drag-and-drop example)
 
@@ -160,7 +159,7 @@ htmx.ajax("GET", `/htmx/nav/r/${window.ctx.getRepoId()}`, { swap: "innerHTML", t
 | File | Purpose |
 |---|---|
 | `static/js/constants.js` | All string constants: events, attributes, store keys, view names |
-| `static/js/app.js` | Alpine init, store registration, `initApp()` entry point |
+| `static/js/app.js` | thin bootstrap that exposes `initApp()` and starts `FrontendApp` |
 | `static/js/context.js` | `window.ctx` — repo ID and metadata field settings |
 | `static/js/models/folder.js` | DOM wrapper around folder tree nodes |
 | `static/js/common/modal.js` | `showModal`, `showAssetDetailModal`, `closeModal` |
