+++ altitude/views/htmx/people.scala.html
@@ -170,6 +170,5 @@
     window.ctx.setRepoId("@{ RequestContext.getRepository.persistedId }")
 </script>
 
-<script type="module" src="/static/js/people/dragon-drop.js"></script>
