+++ altitude/views/htmx/person.scala.html
@@ -73,30 +73,5 @@
     @htmx.html.person_inner(person)
 </div>
 
-<script type="module">
-    import { Const } from "/static/js/constants.js"
-
-    const badMatchEl = htmx.find("#markAsBadMatch")
-
-    if (badMatchEl) {
-        badMatchEl.addEventListener("htmx:afterRequest", function (evt) {
-            if (evt.detail.successful === false) {
-                return
-            }
-
-            const personMarkedAsBadMatchEvent = new CustomEvent(
-                Const.events.personMarkedAsBadMatch,
-                {
-                    detail: {
-                        "@{ Api.Field.PERSON_ID }": "@{ person.persistedId }",
-                    },
-                },
-            )
-            document.body.dispatchEvent(personMarkedAsBadMatchEvent)
-        })
-    }
-</script>
-
-<script type="module" src="/static/js/person/dragon-drop.js"></script>
